requirements.txt
elasticsearch>=7.16.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
flake8==4.0.1
Flask
future-annotations==1.0.0
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.22.2
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pygeoapi
pyparsing==3.0.6
pyrsistent==0.18.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rdflib==6.0.2
requests==2.26.0
Shapely==1.8.0
six==1.16.0
tinydb==4.5.2
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.7
Werkzeug==2.0.2
